backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
polars==0.20.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
elasticsearch==8.11.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rerankers==0.4.0